Rejig the user guidance
As we add pages, the guidance was becoming a bit awkward to navigate.
I've removed index pages per folder, in favour of one main index page.
This makes it easier for a user to scroll and see all the content at
once (you had to go in and out of index pages before to see content).  I've reordered it so that the user guidance and concepts are at
the top, and our team guidance is at the bottom.  I've also added a
getting help page.  This format is similar to the cloud platform format
and I think clearer, but we should reassess as we get user feedback and
change accordingly.