Add file check tests for opa
OPA doesn't check the presence/absence of files because to test the json
you pass in a file, if you don't pass in a file then it can't run a test
against it.  To get round this I've added a bash check, reusing some of
the original bash check jq where possible.

Also added missing performance hub files to enable to tests to be
passing.