Generate environments repo codeowners file
We want to prevent other teams from amending directories which are not
their application. This creates the automation to do this with code
owners.  Note that the github teams are assigned per environment, and
the code is stored per application (which then uses workspaces to change
environment). This means there may be multiple teams per applicaion.
Here we are giving any team with access to an environment, the access to approve the code.