Change to member-providers file
To secure the members terraform, we needed to change the core-vpc role
to use a business unit scoped role rather than an all access role, and
remove the role for shared services. To do this via sed/file
manipulation was becoming messy, so instead we've created a separate
member-providers.tf template file, and depending on whether the
provision member directories script or the provision environment
directories script is running the correct template will be copied. This
also allows us to further modify the repective providers.tf files in
future if we need to and not have to further modify the scripts.