Add role to allow dns record modification
Because of a 64 charater limit on the domain name on a public
certificate, we've had to change the domain name to be
modernisation-platform.service.gov.uk, and have the SAN as eg
platforms-development.modernisation-platform.service...

https://docs.aws.amazon.com/acm/latest/APIReference/API_RequestCertificate.html#ACM-RequestCertificate-request-DomainName

By doing this it means when a user creates a certificate it needs to be validated from
the modernisation-platform domain level. This is created in the
core-network-services account so we need to give user pipelines a role
to be able to create these records without being able to do anything
else in this account.

Permission set taken from here:
https://docs.aws.amazon.com/Route53/latest/DeveloperGuide/r53-api-permissions-ref.html#required-permissions-resource-record-sets