Add team access to environments repo
Adding teams that need access to the environment repo. We need to
explicitly specify team access in the repo for github environments
approvals to work (as we require a team approval, that team must have
access).  Hopefully
https://github.com/ministryofjustice/operations-engineering/issues/139
will be possible and we can just add the parent team once here.  If not
we will need to automate the updating of this list of application teams
and trigger the github workflow to run when any changes to environment
json files are made.