Enable dynamic tagging
Currently tags are all statically defined in the locals file for member
accounts. This change pulls the tags from the environments file, so if
the tags in the environment json file are changed this will be reflected
the next time the infrastrucuture terraform for that account is run.

I am getting the .json file directly from github over http, rather than
referencing the file in the repository, because the same template is
used in the environments repository, and this will work from both
repositories to pull the file to generate the tags. This keeps it
consistent accross the repos.

Also some minor updates to the tags to add things like the source code
repository. This will help going forward to identify if a resource in an
account was generated from the members environments repo or this repo.