Add slack notification for workflow failure
This adds a notification to slack when a workflow fails. Note this is
added only to jobs that run when merged into main.  There is no point in
added alerts to jobs that run when a PR is created - we'll see if it
fails on the PR.  But currently once we merge a PR and a job runs you
either have to manually check that the workflow ran ok, or assume it
runs sucessfully.  As 99% of the time they run sucessfully it has caught
us out in the past where things have failed and we've not seen the
impact of the failure until later.