Fix environments repo workspace creation failing
Fixes 920

This is because the script is creating the temporary files for the main repo, running init and creating the workspace, then changing the backend and running init and create again for the environments repo. Because init has been run once before it now thinks there is a change in the backend. We need to clear the temp folders and start from scratch again for the environments repo.

The reason it works on the 2nd time is because it doesn't create the first workspace and then switch.

This fixs puts all the file creation code into a new function, and calls
that function before it is needed. If any thing already exists it is
deleted in the function before recreating the files.