Remove format-code required
The code formatter hangs when it has to make a commit because it uses
the same repo token therefore doesn't trigger the workflow to run after
making a commit. This leaves the PR in a hung state as it can't be
merged until the required check has run.

https://github.community/t/push-from-action-does-not-trigger-subsequent-action/16854/5

Removing the need for the formatter to be required, it is only
formatting code and not critical that it runs.