Skip baseline for all anlytical platform accounts
Widening the criteria for skipping the scheduled baseline to include all
analytical platform accounts.  The are deleting their default VPC which
causes the baseline to fail as we expect to manage them in terraform. We
do want default VPCs to be deleted but we can't currently do this until
v4 of the AWS provider is released.  Keeping this work around for now
rather than forcing AP to have default VPCs or removing the default vpc
restrictions from the baseline which would impact all other accounts.