Fix bug where the app json file is needed
The application json file is part of the templates and needed for the
environments repository, but it isn't needed here.  Commenting rather
than removing as it's less code in the provisioning script to comment
rather than remove. Also leaving as a comment also serves as a reminder
as to how we should handle environment specific variables if it is ever
needed in this repository.