Add regional replication to S3 state bucket
Adding regional replication and switching to customer managed KMS keys for the
terraform state bucket. This follows the AWS best practice guidelines,
and if there is ever an issue where we loose access to the London region
we will still be able to see the last known state of the platform.