Fix permissions with perf hub ECR repo
The existing permissions would only allow the deployment users and the
main modernisation platform account to push and pull from the repo. But
the actual environment account will be the account pulling the image
when ECS starts the container in that account.  This change splits out
the push and pull permissions in the module to make it possible to have
different permissions for each.