Fix cyclic dependancy in secrets
This was working locally but when running on the pipeline, because the
pipeline assumes a role which references the environment management
secret it causes a cyclic dependancy.  Added the secret section back in
and it will need to be commented out as well to run locally.  Updated
the guidance to reflect this.