Store the slack url as a secret
The slack url currently gets overwritten every time we deploy the github
terraform, by creating the secret in AWS (and setting it manually via
the console) next time the github code runs it will pull from the secret
when there is a new version and only update it then.  There is then no
need to set this manually in github.