After enabling security hub alarms for the modernisation platform core account, there were a lot of unauthorized api alarms -
```
Sum unauthorised-api-calls GreaterThanOrEqualToThreshold 1.0
```

Tracking down the alarm reason in cloudtrail shows that the unauthorised
call was to DescribeAccountAttributes from the member-ci user.

This is because calls from the member-ci user assume a role normally,
and the only thing the ci user has direct permissions for is to access
the state. When the Terraform AWS provider first starts it checks the
supported EC2 platforms which uses the above permission.  We don't need
it to do this for the non member accounts that the member-ci has access
to, as it will only ever create EC2 resources in the member account
(which has this permission as it assumes the member admin role)

https://registry.terraform.io/providers/hashicorp/aws/latest/docs#skip_get_ec2_platforms

So this PR is adding the skip option for all the AWS providers that are
not the members account, eg the modernisation-platform core, network-services
and core-vpc accounts to the template for member accounts.