Add temporary read only users to core accounts
This is for an IT health check.  This adds the testers to the
modernisation platform core account and gives them read only access to
that account and the ability to assume a readonly role in the other core
accounts.

The readonly role is created in the required accounts.

This PR should be reverted and all users and roles removed after the
health check.