Add automation to create collaborator IAM users
Closes #158

This adds functionality to add external collaborators who are not part
of the ministryofjustice GitHub organisation to the Modernisation
Platform. See the user guidance added in this PR for the full process to
set up users.

Overview:

- collaborators.json - json file where all collaborators and their
account associations can be added
- OPA tests to ensure the above json is in the correct format
- new roles added to all core and member accounts via delegate access -
read-only and developer, collaborators will only be able to assume these
roles for their listed accounts
- new collaborators module to set up users
- new IAM group and policy for collaborator users
- call the new module from the modernisation-platform core account
- minor change to github create environments script (we are not amending
this script to add collaborators but a small fix was noticed)