Remove automation for terraform in unrestricted accounts
We initially set up the modernisation-platform-environments repo to
include automation for generating terraform for the unrestricted member
accounts. However as time goes on and we are applying updates to code,
we are finding that no one is using this code, and also as they have
full access to the accounts, they are able to do things that break the
infrastructure that we're trying to automate.

Therefore we have decided to remove all member automation for the
unrestricted accounts, these will be soley managed out side of the
platform.  Our efforts will be focused on moving these to member
accounts if that is sensible rather than trying to do a hybrid approach.