Trigger core-shared-services workflow on new ac
The core shared services account creates a key per business unit which
is shared to the accounts for that business unit. The TF pulls these
accounts from the environments secret, but this only happens when the TF
is run. This will ensure when new accounts are added to the
modernisation platform, the core-shared-services workflow runs again
getting the latest account numbers.