Add backup information
Added a new disaster recovery section and information about the backups
that we take.  We can also add runbooks and other DR information here.

Added a new section for how to guides, the getting started section was
getting crowded with things that not all users needed.  By splitting
this out we can have a more details list of guides explaining how to do
different things and keep the getting started for the basics.