Add issues to all repositories
Currently we are failing to meet [MoJ repository
standards](https://github.com/ministryofjustice/github-repository-standards) for the
majority of our repositories because we do not have issues enabled on
them.  This was intentional so that all issue could be on our main repo
and we could manage them using github projects.  Now we are using an
org level project, we can add issues from any repo to it. Following this
change we can add this workflow to each repo to automatically add any
issues to the project from any repo to it. Following this change we can
add [this
workflow](https://github.com/ministryofjustice/modernisation-platform/blob/main/.github/workflows/add-issues-to-project.yml) to each repo to automatically add any issues to the
project.

Also removed a repository that we will not be using, our hello world
application will sit in the modernisation-platform-environments repo.