Modify targeted plan to remove nacls module
The following error:

```
Error: Invalid for_each argument

  on ../../modules/vpc-nacls/main.tf line 32, in resource "aws_network_acl_rule" "open_endpoint_cidrs_for_data_subnets_egress":
  32:   for_each = {
  33:     for key, data in var.cidrs_for_s3_endpoints :
  34:     "${data.name}-${key}" => data
  35:   }
    ├────────────────
    │ var.cidrs_for_s3_endpoints is a list of dynamic, known only after apply
```

Comes about as the vpc needs to be created before the data look-up is
valid.  We have come across this before and resolved it by running a
targetted plan, then a targetted apply to create the vpc resources then
a full apply to build out the rest.  Following this same pattern to
remove the nacl module from the targetted plan so it doesn't fail.

Adding as a risk on the register that we are not seeing the full plan
when approving these PRs.