Fix failing local plans for dns
```
Error finding Route 53 Hosted Zone: AccessDenied: User: arn:aws:sts:::assumed-role/read-dns-records/aws-go-sdk is not authorized to perform: route53:ListHostedZones because no identity-based policy allows the route53:ListHostedZones action
```

The above error was occuring on local plans, this is due to the addition
of application hosted zones not being added to the policy.

This changes to give get and list permissions to everything, and for the
write role add in the application zones.