Add testing-ci user and give required permissions
This creates a testing-ci user in the testing-test account.  The current
testing user is created in the landingzone account, but this means that
it doesn't have access to terratest aws module to make additional aws
tests outside of terraform.

This means that testing pipelines will need to be updated to use the new
testing user credentials, and when running locally tests these
credentials will need to be used as well.  Since they are stored in
secret manager a simple script can be written to easily get them.  These
changes will be made in subsequent PRs.