Switch module repos to use new testing-ci creds
This changes the credentials that are used in the terraform module repos
to run the terratests from the old testing-ci user in the modernisation
platform account to the new testing-ci user in the testing-test account.
This was needed due to terratest not being able to assume roles during a
test.