Replace security hub findings script with Go one
The existing script does not work and is not fit for purpose. This
replaces that script with a new Go script that retrieves the findings
across all MP accounts.  Note that this only gets findings from the
security hub product which are CRITICAL or HIGH.

Adding in other
products such as Config and Inspector findings can be done later, but
cause and issue when generally adding as they don't contain the same
fields, leading to nil pointer exceptions.