Widen some permissions to write
Terraform documenter failed when writing changes to a PR when contents
was set to read.  Assuming that other things that need to write to a PR
also need write permission, adding this in.  After everything is
confirmed as working as is, we can then start changing things to read
and see if they break.