Change order of new environment workflows
The order of these were changed to fix a bug here - https://github.com/ministryofjustice/modernisation-platform/pull/2855

But this then brakes the SSO as the role doesn't exist yet.  There is no
way round this dependancy, so reversing this change and will have to
find another way to fix the bug.