Use GitHub Actions matrix for sceduled baselines
The baselines workflow is taking over an hour, by switching to a matrix
the jobs can run concurrently, massively reducing the amount of time
needed.

Note that the limit for using a matrix is 256, we currently are on 62
workspaces so this should last us a while and we can refactor if we
reach the limit.