Remove exit 0 from RAM account script
After the first if statement was successful, the script was exiting with
0 status without running the 2nd if statement which is required if there
  are changes with the networking vars.

The exit 0 statements were put in there to try to fix an error where the
script was running successfully but returning a non zero code and the
workflow was failing.

Removing the set e to try to resolve this.  We will have to see how the
next new account run looks to see if this has resolved it.