Fix network file creation bug
Fixes new networking files created for OAS when they shouldn't be

Previously contains was picking up oasys for oas as it contains oas

By switching to test() we can add the start of line ^ and the - dash
before the environment line.

So this will work for:

oas-development
oasys-development

We might stil have an issue if something came along with the name

oas-sys-development

But let's hope that doesn't happen.