Turn off strict checking
This means branches don't have to be updated against main prior to
merging.  This is off by default and only comes on when you have status
checks, but it is very annoying and doesn't add much value as merge
conflict will still prevent bad merges and the opa tests are rarely
updated and protected by codeowners.