Add generic services to raise incidents manually
Specific areas such as TGW or DDOS have their own services, but if there
is a generic issue which isn't covered by one of the other services we
need to be able to raise it somewhere.

Following the incident meeting we decided we preferred to keep
networking, operations and security as business services as it's makes
it clear to users where the problem is, but add additional general
services which feed in to these business services.