Refactor to names not CIDR ranges
This moves the definition of the CIDR ranges in to one file, and then
the firewall files can use words rather than ranges to define the
firewall rules.

This helps in a number of ways -

1. Easier to read and create
2. If CIDR ranges change we only update them in one place
3. We automatically get all of the MP ranges for the general subnet sets
   so no need to look them up.

This PR is an initial draft to demo this.  Other firewall files will be
added if it is agreed to move forward.

Note the plan shows no changes.