SCA ignores for open policies
These are mostly for policies which are directly attached to the
resource or similar and there for resources needs to be *.  They have
been ignored for tfsec before but this is a new check with checkov now
doing the same thing.  Also a couple of other ignores which have been
verified as ok.