Fix change of github teams not going to approvers
Fixes
https://github.com/ministryofjustice/modernisation-platform/issues/3767.

Github environments were being created in the environments repository
after the application files were being created.  However adding teams to
deployment environment approvers isn't dependant on this.  It is
dependant however on adding the github team to the repository, which
happens in the terraform-github.yml workflow, so moving this job to
there.