Allocate /21 CIDR range to data platforming
This range will be used for the data platforming team to build a VPC for
Airflow and any other non platform networking requirements.

Note this range is not used in the Modernisation Platform core accounts
or core-vpc accounts and management of the range is delegated to the
data platforming team.