Change SSM resource sync to be opt in
Currently the bucket policy and KMS policy in the organisation security
account SSM resource sync bucket require account numbers to be added in
order to restrict access to the organisation, as you can't use org
principal paths with service principals (ssm in this case).  This has
been raised with AWS and a feature request added.  Until then our
options are limited as we can't create the resource sync until the org
sec terraform has run after the new account is created and then the
account has permissions to the bucket.

Triggering the org sec to run after the env is created but before the
bootstrap runs is a possibility, but not at this point as the org sec tf
is run manually.  This looks to be the only way to ensure that this
pipeline doesn't fail because of this, but it does mean there is an
extra step for users wishing to use the Oracle licensing solution.  That
will be added to the user guidance when it is created.