Update runbook.html.md.erb
Added line break between standard hours and out of hours sentences. Also changed the text to say that the way to get in touch out of hours is either by email or raising a call on PagerDuty. Removed the bit about criticality - it's the same method regardless and it previously read like you'd get a different type of response if you raised it directly through PD.